+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed bug in different `yearmonth()` underlying data for character inputs. (#226)
 * Fixed C stack error when index contains NA with `as_tsibble()`. (#229)
+* Fixed bug when `by` is missing in `year*()`. (#228)
+* Removed rolling functions.
 
 # tsibble 0.9.3